new API works for serial and mass; still needs implementation for MIDI, HID and x360
