+++ USBHID.h
@@ -24,7 +24,7 @@
 #include <boards.h>
 #include "Stream.h"
 #include "usb_hid.h"
-#include "usb_serial.h"
+#include <CompositeSerial.h>
 
 #define USB_HID_MAX_PRODUCT_LENGTH 32
 #define USB_HID_MAX_MANUFACTURER_LENGTH 32
@@ -718,40 +718,6 @@ public:
     }
 };
 
-class USBCompositeSerial : public Stream {
-public:
-    USBCompositeSerial (void);
-
-    void begin(void);
-
-	// Roger Clark. Added dummy function so that existing Arduino sketches which specify baud rate will compile.
-	void begin(unsigned long);
-	void begin(unsigned long, uint8_t);
-    void end(void);
-
-	operator bool() { return true; } // Roger Clark. This is needed because in cardinfo.ino it does if (!Serial) . It seems to be a work around for the Leonardo that we needed to implement just to be compliant with the API
-
-    virtual int available(void);// Changed to virtual
-
-    uint32 read(uint8 * buf, uint32 len);
-   // uint8  read(void);
-
-	// Roger Clark. added functions to support Arduino 1.0 API
-    virtual int peek(void);
-    virtual int read(void);
-    int availableForWrite(void);
-    virtual void flush(void);	
-	
-    size_t write(uint8);
-    size_t write(const char *str);
-    size_t write(const uint8*, uint32);
-
-    uint8 getRTS();
-    uint8 getDTR();
-    uint8 isConnected();
-    uint8 pending();
-};
-
 extern HIDMouse Mouse;
 extern HIDKeyboard Keyboard;
 extern HIDJoystick Joystick;
@@ -776,3 +742,4 @@ extern const HIDReportDescriptor* hidReportBootKeyboard;
 
 #endif
         
+
