+++ usb_hid_device.cpp
@@ -32,11 +32,6 @@
  * USB HID interface
  */
 
-#if defined(SERIAL_USB)
-static void rxHook(unsigned, void*);
-static void ifaceSetupHook(unsigned, void*);
-#endif
-
 #define USB_TIMEOUT 50
 
 USBHIDDevice::USBHIDDevice(void) {
@@ -133,13 +128,6 @@ void USBHIDDevice::begin(const uint8_t* report_descriptor, uint16_t report_descr
         usb_generic_set_parts(parts, numParts);
         usb_generic_enable();        
             
-#if defined(SERIAL_USB)
-        if (serialSupport) {
-            composite_cdcacm_set_hooks(USBHID_CDCACM_HOOK_RX, rxHook);
-            composite_cdcacm_set_hooks(USBHID_CDCACM_HOOK_IFACE_SETUP, ifaceSetupHook);
-        }
-#endif
-            
 		enabled = true;
 	}
 }
@@ -226,112 +214,5 @@ uint16_t HIDReporter::getOutput(uint8_t* out, uint8_t poll) {
     return usb_hid_get_data(HID_REPORT_TYPE_OUTPUT, reportID, out, poll);
 }
 
-#if defined(SERIAL_USB)
-
-enum reset_state_t {
-    DTR_UNSET,
-    DTR_HIGH,
-    DTR_NEGEDGE,
-    DTR_LOW
-};
-
-static reset_state_t reset_state = DTR_UNSET;
-
-static void ifaceSetupHook(unsigned hook, void *requestvp) {
-    (void)hook;
-    uint8 request = *(uint8*)requestvp;
-
-        // Ignore requests we're not interested in.
-    if (request != USBHID_CDCACM_SET_CONTROL_LINE_STATE) {
-        return;
-    }
-
-    // We need to see a negative edge on DTR before we start looking
-    // for the in-band magic reset byte sequence.
-    uint8 dtr = composite_cdcacm_get_dtr();
-    switch (reset_state) {
-    case DTR_UNSET:
-        reset_state = dtr ? DTR_HIGH : DTR_LOW;
-        break;
-    case DTR_HIGH:
-        reset_state = dtr ? DTR_HIGH : DTR_NEGEDGE;
-        break;
-    case DTR_NEGEDGE:
-        reset_state = dtr ? DTR_HIGH : DTR_LOW;
-        break;
-    case DTR_LOW:
-        reset_state = dtr ? DTR_HIGH : DTR_LOW;
-        break;
-    }
-
-	if ((composite_cdcacm_get_baud() == 1200) && (reset_state == DTR_NEGEDGE)) {
-		iwdg_init(IWDG_PRE_4, 10);
-		while (1);
-	}
-}
-
-#define RESET_DELAY 100000
-static void wait_reset(void) {
-  delay_us(RESET_DELAY);
-  nvic_sys_reset();
-}
-
-#define STACK_TOP 0x20000800
-#define EXC_RETURN 0xFFFFFFF9
-#define DEFAULT_CPSR 0x61000000
-static void rxHook(unsigned hook, void *ignored) {
-    (void)hook;
-    (void)ignored;
-    /* FIXME this is mad buggy; we need a new reset sequence. E.g. NAK
-     * after each RX means you can't reset if any bytes are waiting. */
-    if (reset_state == DTR_NEGEDGE) {
-        reset_state = DTR_LOW;
-
-        if (composite_cdcacm_data_available() >= 4) {
-            // The magic reset sequence is "1EAF".
-            static const uint8 magic[4] = {'1', 'E', 'A', 'F'};	
-
-            uint8 chkBuf[4];
-
-            // Peek at the waiting bytes, looking for reset sequence,
-            // bailing on mismatch.
-            composite_cdcacm_peek_ex(chkBuf, composite_cdcacm_data_available() - 4, 4);
-            for (unsigned i = 0; i < sizeof(magic); i++) {
-                if (chkBuf[i] != magic[i]) {
-                    return;
-                }
-            }
-
-            // Got the magic sequence -> reset, presumably into the bootloader.
-            // Return address is wait_reset, but we must set the thumb bit.
-            uintptr_t target = (uintptr_t)wait_reset | 0x1;
-            asm volatile("mov r0, %[stack_top]      \n\t" // Reset stack
-                         "mov sp, r0                \n\t"
-                         "mov r0, #1                \n\t"
-                         "mov r1, %[target_addr]    \n\t"
-                         "mov r2, %[cpsr]           \n\t"
-                         "push {r2}                 \n\t" // Fake xPSR
-                         "push {r1}                 \n\t" // PC target addr
-                         "push {r0}                 \n\t" // Fake LR
-                         "push {r0}                 \n\t" // Fake R12
-                         "push {r0}                 \n\t" // Fake R3
-                         "push {r0}                 \n\t" // Fake R2
-                         "push {r0}                 \n\t" // Fake R1
-                         "push {r0}                 \n\t" // Fake R0
-                         "mov lr, %[exc_return]     \n\t"
-                         "bx lr"
-                         :
-                         : [stack_top] "r" (STACK_TOP),
-                           [target_addr] "r" (target),
-                           [exc_return] "r" (EXC_RETURN),
-                           [cpsr] "r" (DEFAULT_CPSR)
-                         : "r0", "r1", "r2");
-
-            /* Can't happen. */
-            ASSERT_FAULT(0);
-        }
-    }
-}
-#endif
 
 USBHIDDevice USBHID;
